+++ src/natcap/invest/ndr/ndr.py
@@ -55,6 +55,7 @@ MODEL_SPEC = spec.ModelSpec(
         ["flow_dir_algorithm", "threshold_flow_accumulation",
          "k_param", "runoff_proxy_av"]
     ],
+    aoi_input_id="watersheds_path",
     inputs=[
         spec.WORKSPACE,
         spec.SUFFIX,