+++ src/natcap/invest/urban_mental_health/urban_mental_health.py
@@ -69,6 +69,7 @@ MODEL_SPEC = spec.ModelSpec(
         ["lulc_base", "lulc_alt", "lulc_attr_csv"],
         ["target_projection_id", "target_pixelsize_id"]
     ],
+    aoi_input_id="aoi_path",
     default_projection_id="aoi_path",
     inputs=[
         spec.WORKSPACE,