+++ workbench/src/renderer/components/InvestTab/index.jsx
@@ -325,6 +325,7 @@ class InvestTab extends React.Component {
                     modelID={modelID}
                     argsSpec={argsSpec}
                     inputFieldOrder={modelSpec.input_field_order}
+                    aoiInputId={modelSpec.aoi_input_id || ''}
                     defaultProjectionId={modelSpec.default_projection_id}
                     defaultPixelsizeId={modelSpec.default_pixelsize_id}
                     argsInitValues={argsValues}